QueryBridg, ChatGPT Connect/requirements.txt
Flask==2.0.1
requests==2.25.1
boto3==1.17.72
python-dotenv==0.17.1
